Demos/ABS_Demo/ABS/requirements.txt
python-can>=4.2
gpiod==2.2.3
